+++ CHANGELOG.md
@@ -103,3 +103,12 @@ New :
 {content}
 ```
 Fix the issue that poll thread HTML content is not parsed correctly.
+
+# 1.1.10
+Add two link builder in YamiboRoute :
+```kotlin notebook
+//Reply a specific post(mention a post)
+data class PostReply(val threadId: ThreadId, val postId: PostId, val page: Int = 1)
+//Reply the thread : 
+data class ThreadReply(val threadId: ThreadId, val page: Int = 1)
+```